Migrate westchester (Election Book PDF, positional party-code row)
